refactor: replaceNewLines 분리
