+++ src/hooks/newLine.tsx
@@ -1,0 +1,5 @@
+export const replaceNewLines = (text: string): string => {
+  return text.replace(/\r?\n/g, '\n');
+};
+
+export default replaceNewLines;
